refactor: rename EventTypeDialog to CalendarDialog and update related imports; implement EventManagement component with event handling logic
