+++ src/app/(calendar)/page.tsx
@@ -17,7 +17,7 @@ import {
   useUpdateCalendar,
   useCreateCalendar,
 } from "@/hooks/calendar/use.calendar";
-import { EventTypeDialog } from "@/components/events/event-type-dialog";
+import { EventTypeDialog } from "@/components/events/calendar-dialog";
 import { useCalendarDialog } from "@/context/calendar-dialog-context";
 
 export default function CalendarPage({
@@ -35,11 +35,7 @@ export default function CalendarPage({
 }) {
   const [currentView, setCurrentView] = useState<CalendarView>("day");
   const { deleteEvent } = useDeleteEvent();
-  const {
-    updateEvent: updateEventAPI,
-    isUpdatingEvent,
-    updateEventError,
-  } = useUpdateEvent();
+  const { updateEvent: updateEventAPI, updateEventError } = useUpdateEvent();
 
   const { updateCalendar } = useUpdateCalendar();
   const { createCalendar } = useCreateCalendar();
@@ -116,7 +112,6 @@ export default function CalendarPage({
   const handleUpdateEvent = (eventData: Partial<Event>) => {
     if (!selectedEvent) return;
 
-    // Lưu trạng thái cũ để rollback nếu cần
     const originalEvent = events?.find((e) => e.id === selectedEvent.id);
     if (originalEvent) {
       setOptimisticUpdates(
@@ -124,7 +119,6 @@ export default function CalendarPage({
       );
     }
 
-    // Optimistic update cho dialog updates
     setEvents((prevEvents) => {
       if (!prevEvents) return prevEvents;
       return prevEvents.map((event) =>
@@ -132,7 +126,6 @@ export default function CalendarPage({
       );
     });
 
-    // Gọi API để cập nhật backend
     updateEventAPI({
       id: selectedEvent.id,
       data: {
@@ -144,7 +137,6 @@ export default function CalendarPage({
     setSelectedEvent(null);
   };
 
-  // Rollback optimistic updates nếu có lỗi
   useEffect(() => {
     if (updateEventError) {
       setEvents((prevEvents) => {
@@ -168,8 +160,6 @@ export default function CalendarPage({
   };
 
   const handleOptimisticUpdate = (eventId: string, updatedEvent: Event) => {
-    // Chỉ cập nhật state cho drag/resize operations
-    // Dialog updates sẽ được xử lý bởi React Query
     setEvents((prevEvents) => {
       if (!prevEvents) return prevEvents;
       return prevEvents.map((event) =>
