+++ src/app/globals.css
@@ -143,5 +143,5 @@ html {
   @apply border border-gray-300 focus:border-blue-500 focus:ring-blue-500;
 }
 .decored-input {
-  @apply border border-gray-300 focus:border-blue-500 focus-visible:ring-blue-500 dark:selection:bg-[#658DBD] selection:bg-blue-500;
+  @apply border border-gray-300 focus:border-blue-500 focus-visible:ring-blue-500 dark:selection:bg-[#658DBD] selection:bg-blue-500 selection:text-white;
 }
